+++ Win32Window.cpp
@@ -67,10 +67,10 @@ Win32Window::Win32Window(const std::string& title, int sizeX, int sizeY, bool fu
 	WINDOWCLASS,    // name of the window class
 	title.c_str(),   // title of the window
 	fullScreen ? WS_POPUP|WS_VISIBLE : WS_OVERLAPPEDWINDOW|WS_POPUP|WS_VISIBLE|WS_SYSMENU|WS_MAXIMIZEBOX|WS_MINIMIZEBOX,    // window style
-						(int)position.x,	// x-position of the window
-                        (int)position.y,	// y-position of the window
-                        (int)size.x,		// width of the window
-                        (int)size.y,		// height of the window
+						position.x,			// x-position of the window
+                        position.y,			// y-position of the window
+                        size.x,				// width of the window
+                        size.y,				// height of the window
                         NULL,				// No parent window!
                         NULL,				// No Menus!
 						windowInstance,		// application handle
